requirements.txt
# Requirements

os # Genaral package
smtplib # Genaral package
time # Genaral package
python-dotenv # Local package